requirements.txt
numpy>=1.24.0
scipy>=1.10.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
